requirements/Base.txt
amqp==5.0.9
asgiref==3.4.1
attrs==19.3.0
Automat==0.8.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blinker==1.4
celery==5.2.3
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.12
django-allauth==0.47.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-crispy-forms==1.13.0
django-embed-video==1.4.1
django-environ==0.8.1
django-modeltranslation==0.17.3
django-registration-redux==2.9
django-timezone-field==4.2.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.4.2
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==21.3
pexpect==4.6.0
Pillow==8.4.0
platformdirs==2.4.1
prompt-toolkit==3.0.26
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.6
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.5
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
redis==4.1.2
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.2
SQLAlchemy==1.4.29
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==27.5
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.0.0
virtualenv==20.13.0
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.13.3
zipp==1.0.0
zope.interface==4.7.1
